refactor: move onboarding into own component
