+++ src/components/TeacherViews/Onboarding/Onboarding.view.tsx
@@ -1,0 +1,98 @@
+import { zodResolver } from "@hookform/resolvers/zod";
+import {
+  OakFlex,
+  OakHeading,
+  OakPrimaryButton,
+} from "@oaknational/oak-components";
+import { useCallback, useEffect } from "react";
+import { useForm } from "react-hook-form";
+import { z } from "zod";
+
+import useSchoolPicker from "@/components/TeacherComponents/ResourcePageSchoolPicker/useSchoolPicker";
+import ResourcePageSchoolPicker from "@/components/TeacherComponents/ResourcePageSchoolPicker";
+
+const onboardingFormSchema = z.object({
+  school: z
+    .string({
+      errorMap: () => ({
+        message:
+          "Select school, type ‘homeschool’ or tick ‘My school isn’t listed’",
+      }),
+    })
+    .min(
+      1,
+      "Select school, type ‘homeschool’ or tick ‘My school isn’t listed’",
+    ),
+  schoolName: z.string().optional(),
+});
+type OnboardingFormValues = z.infer<typeof onboardingFormSchema>;
+type OnboardingFormProps = OnboardingFormValues & {
+  onSubmit: (values: OnboardingFormValues) => Promise<string | void>;
+};
+
+export const OnboardingView = () => {
+  const { formState, setValue, handleSubmit } = useForm<OnboardingFormProps>({
+    resolver: zodResolver(onboardingFormSchema),
+    mode: "onBlur",
+  });
+
+  const setSchool = useCallback(
+    (value: string, name: string) => {
+      setValue("school", value, {
+        shouldValidate: true,
+      });
+      setValue("schoolName", name, {
+        shouldValidate: true,
+      });
+    },
+    [setValue],
+  );
+
+  const {
+    selectedSchool,
+    setSelectedSchool,
+    schoolPickerInputValue,
+    setSchoolPickerInputValue,
+    schools,
+  } = useSchoolPicker();
+
+  useEffect(() => {
+    if (selectedSchool && schoolPickerInputValue !== "") {
+      setSchool(selectedSchool.toString(), schoolPickerInputValue);
+    }
+  }, [selectedSchool, schoolPickerInputValue, setSchool]);
+
+  const onFormSubmit = async (data: OnboardingFormProps) => {
+    // TODO: something with this data
+    console.log("onboarding form values", data);
+  };
+
+  return (
+    <OakFlex
+      $flexDirection="column"
+      $gap="all-spacing-10"
+      $mv="space-between-m"
+    >
+      <OakHeading tag="h1">Onboarding</OakHeading>
+      <ResourcePageSchoolPicker
+        hasError={formState.errors?.school !== undefined}
+        schoolPickerInputValue={schoolPickerInputValue}
+        setSchoolPickerInputValue={setSchoolPickerInputValue}
+        schools={schools}
+        label={"School"}
+        setSelectedSchool={setSelectedSchool}
+        required={true}
+      />
+      <OakPrimaryButton
+        onClick={
+          (event) => void handleSubmit(onFormSubmit)(event) // https://github.com/orgs/react-hook-form/discussions/8622}
+        }
+        disabled={formState.errors?.school !== undefined || !formState.isValid}
+      >
+        Continue
+      </OakPrimaryButton>
+    </OakFlex>
+  );
+};
+
+export default OnboardingView;
